requirements.txt
# Schwab API wrapper
schwab-py==0.2.1

# Data processing and analysis
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1

# Web requests and APIs
requests==2.31.0
aiohttp==3.9.1

# LLM API clients
anthropic==0.16.0
openai==1.12.0
google-generativeai==0.3.1  # Google Gemini API

# Financial analysis and market data
yfinance==0.2.35
ta==0.11.0
stockstats==0.5.4
pandas-ta==0.3.14b0
polygon-api-client==1.12.3
finnhub-python==2.4.18
alpha_vantage==2.3.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
rich==13.7.0
schedule==1.2.1
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Development
black==23.12.0
isort==5.13.2
flake8==6.1.0